Construction lands where the cursor is, and a room's cast survives travel
The review of 7c6afa5 found a document-format inconsistency and a live
placement regression; this is the fix wave the controller's rulings called
for, behaviour and docs together.

The ground keeps whole levels. Codex's height field wrote a renderer constant
(0.35) into saved content to reach a fractional Z, mixing two units in one
array while Raise and Lower still stepped by one. That half is reverted -
Int16Array, an integer schema, no paintTerrainAtHeight and no "Apply Z height
when painting ground" - and the construction half is kept as it was, in tile
units. Part 2's spec picks the single unit; until then DEVELOPING says in so
many words that nothing may convert between them.

Placing a prop, object or creature at ground level raycasts the terrain again
rather than a flat plane at the build level. At this camera pitch the plane
projects a tile raised two levels about two thirds of a tile short of where it
stands, so a click aimed at the vault's plateau landed a row behind it; the
new e2e clicks the plateau through screenOf and reads the tile back out of the
exported project. The plane is still what the build tools use, and what a
pointer off the board or above it falls back to.

Content authored outside the board has one meaning. It is legal - it is how a
street is dressed around a room - so the validator reports one warning per
creature, object or prop out there and skips the impassable, reachability and
shares-a-tile checks that used to call it an error, and both play paths skip
it rather than standing a creature up at no tile at all.

A room's cast is reconciled on every entry, not only on the way back from the
editor. DemoScene remembers the placement ids each scene was last stood up
from, so travelling back into a room brings in what the designer added while
the party was away, takes out what they deleted, and leaves out what a script
removed. main.ts's own bookkeeping goes with it.

Terrain's tab-to-tool mapping is data in modes.ts with tests, and every tool
has a tab: with the Tiles tab open and Erase in hand the rail used to show
neither. The coordinate rules move to scene/building.ts, setZ becomes
controller.setBuildLevel, and the shell's unchecked casts become lookups.

The husk body stays for the 127 SRD creatures with no art, in edit and in
play, but SceneView now records the substitution: missing() still lists the
id and modelSource() answers fallback:husk, so the diagnostic stays honest.

Minor, from the review's list: R and Rotate only while a build tool is in
hand; the camera pins to the build plane only for build tools, so Home works
again with a creature at Z 3.25; the brush preview stops stretching one piece
across the footprint; the Z and go-to controls show only for tools that use
them; keys on three lists; a real tuple out of buildingParts; visibleChunks
counts what the frustum found; the preview does not run in play; clip sets are
dropped with their groups; leaving the editor re-poses tokens instead of
forcing them visible; undo redraws the scenery once; erase reads the document
once per edit rather than once per tile.

Verification: npx tsc --noEmit clean; npx vitest run 1747 tests across 87
files; npx playwright test 97 passed in 3.4 minutes. Each regression test was
recorded failing first: the plateau click landed on y=2, the off-board objects
"shared a tile", the off-board creature was stood up in play, and the creature
authored while the party was elsewhere was missing on return.

+++ src/engine/grid/grid.ts
@@ -40,7 +40,7 @@ export class TileGrid {
   readonly height: number;
   readonly palette: TerrainPalette;
   /** Elevation level per tile, row-major. */
-  readonly heights: Float64Array;
+  readonly heights: Int16Array;
   /** Terrain palette index per tile, row-major. */
   readonly terrain: Uint8Array;
 
@@ -54,7 +54,7 @@ export class TileGrid {
     this.palette = options.palette ?? new TerrainPalette();
 
     const count = width * height;
-    this.heights = new Float64Array(count);
+    this.heights = new Int16Array(count);
     this.terrain = new Uint8Array(count);
     if (options.fillHeight !== undefined && options.fillHeight !== 0) {
       this.heights.fill(options.fillHeight);
